Updates to address code review comments
