+++ udfs/migration/teradata/test_cases.yaml
@@ -22,42 +22,42 @@ last_day:
     expected_output: DATE("2019-04-30")
 nullifzero:
   - test:
-    input: -1
-    expected_output: -1
+    input: CAST(-1 AS INT64)
+    expected_output: CAST(-1 AS INT64)
   - test:
-    input: 0.0
+    input: CAST(0.0 AS FLOAT64)
     expected_output: (NULL)
   - test:
-    input: 0
+    input: CAST(0 AS INT64)
     expected_output: (NULL) # NULL is a type in YAML so wrap it in parenthesis
   - test: # Teradata casts strings into numeric
     input: CAST(0 AS STRING)
     expected_output: (NULL)
   - test:
-    input: 1
-    expected_output: 1
+    input: CAST(1 AS INT64)
+    expected_output: CAST(1 AS INT64)
   - test:
-    input: 1.1
-    expected_output: 1.1
+    input: CAST(1.1 AS FLOAT64)
+    expected_output: CAST(1.1 AS FLOAT64)
   - test:
     input: (NULL)
     expected_output: (NULL)
 nvl:
   - test:
-    input: -1, 1
-    expected_output: -1
+    input: CAST(-1 AS INT64), CAST(1 AS INT64)
+    expected_output: CAST(-1 AS INT64)
   - test:
-    input: 0, 1
-    expected_output: 0
+    input: CAST(0 AS INT64), CAST(1 AS INT64)
+    expected_output: CAST(0 AS INT64)
   - test:
-    input: NULL, 1
-    expected_output: 1
+    input: NULL, CAST(1 AS INT64)
+    expected_output: CAST(1 AS INT64)
   - test:
-    input: 2, 1
-    expected_output: 2
+    input: CAST(2 AS INT64), CAST(1 AS INT64)
+    expected_output: CAST(2 AS INT64)
   - test:
-    input: 3, 1
-    expected_output: 3
+    input: CAST(3 AS INT64), CAST(1 AS INT64)
+    expected_output: CAST(3 AS INT64)
   - test:
-    input: 4, 1
-    expected_output: 4
+    input: CAST(4 AS INT64), CAST(1 AS INT64)
+    expected_output: CAST(4 AS INT64)
